refactor: simplify user authentication logic and enhance password matching                 -vercelとnetlifyではapiでパスワード確認。                 それ以外の環境では直接DBに問い合わせて確認に修正。
