+++ app/api/user/route.ts
@@ -1,25 +1,11 @@
-import bcrypt from "bcryptjs"
-import type { User } from "next-auth"
-import { getUserByName } from "@/app/lib/db/queries/queries"
+import { passwordMatch } from "@/app/lib/auth/passwordMatch"
 
 export async function POST(req: Request) {
   const { name, password } = await req.json()
 
   try {
-    console.log("[API]Start", { name, password: password ? "••••" : password })
-    const result = await getUserByName(name)
-    console.log("[API]DB result", result)
-    const passwordMatch = await bcrypt.compare(password, result.password)
-    console.log("[API]Password match result", passwordMatch)
-    if (passwordMatch) {
-      const user: User = {
-        id: result.id.toString(),
-        name: result.name,
-        email: null,
-        image: null,
-      }
-      return Response.json({ user }, { status: 200 })
-    }
+    const user = await passwordMatch(name, password)
+    return Response.json({ user }, { status: 200 })
   } catch (error) {
     console.error("[API]Error retrieving user", error)
     return Response.json(
@@ -31,5 +17,4 @@ export async function POST(req: Request) {
       { status: 500 },
     )
   }
-  return Response.json({ user: null }, { status: 401 })
 }
