+++ next.config.ts
@@ -1,17 +1,13 @@
 import type { NextConfig } from "next"
-import { webpack } from "next/dist/compiled/webpack/webpack"
 import withRspack from "next-rspack"
 
-const ignorePluginResourceRegExp = /^pg-native$|^cloudflare:sockets$/
 const mp3TestRegExp = /\.(mp3)$/
 
 const nextConfig: NextConfig = {
+  env: {
+    NETLIFY: process.env.NETLIFY ?? undefined,
+  },
   webpack(config) {
-    config.plugins.push(
-      new webpack.IgnorePlugin({
-        resourceRegExp: ignorePluginResourceRegExp,
-      }),
-    )
     config.module.rules.push({
       test: mp3TestRegExp,
       type: "asset/resource",
